增强帖子创建和话题处理逻辑，添加详细错误信息输出；更新作者搜索结果DTO以包含作者ID；修复帖子话题关联查询中的ID类型；改进前端书籍详情和搜索结果页面，增加书籍ID格式转换和加载状态处理。

+++ backend/src/main/java/com/weread/repository/community/PostTopicRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 // 帖子话题关联仓库
 public interface PostTopicRepository extends JpaRepository<PostTopicEntity, PostTopicEntity.PostTopicId> {
     // 根据帖子ID查找所有关联
-    List<PostTopicEntity> findByPostId(Long postId); 
+    List<PostTopicEntity> findByPostId(Integer postId); 
     
     // 根据话题ID查找所有关联
-    List<PostTopicEntity> findByTopicId(Long topicId);
+    List<PostTopicEntity> findByTopicId(Integer topicId);
 }
